moves routers and tests to their own folders
